requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0

# Web scraping and automation
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# Scheduling
APScheduler==3.10.4

# Notifications
python-telegram-bot==20.7

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0
